requirement.txt
Pillow==8.1.0
PyQt5==5.15.0
pyserial
pybind11==2.6.2
opencv-python==4.2.0.32
sympy
